✨ Feat: Use 'DLR_Machine' variable to distinguish
Some parts of my config only work on the DLR machine or are intended to
work on this machine whilst other on my private machine. By introducing
the 'DLR_Machine' variable, I can distinguish and maintain one config
for all machines instead of two branches which have to be merged
sometime.